Allow xdebug to connect back automatically
It was a bit of a puzzle, but I needed these changes to have xdebug connect back properly to PHPStorm when starting a debug session from the PHPStorm interface (i.e. by pressing the green "bug" button). Without this, no breakpoints would seem to hit, unless I enabled listening manually by clicking the icon with a bug and a phone on it.