Allow using absolute paths for sql imports
Currently when you specify a path to an SQL file it is prefixed with `$(PWD)`, so passing `/db/mysite.sql` is translated to `/app//db/mysite-20190902-125000.sql`. The use case is to have all my databases in one directory (e.g. `/db`) and mount that volume in lando. This allows for autocompletion when I start typing `lando db-import /db/mysite-2019`...