Provide documentation on Elasticsearch Connector config
The configuration is a bit unclear on how to get Elasticsearch working with Drupal 8s Elasticsearch Connector. The only reference to connecting to it is the last paragraph on connecting to services externally, which NOT what you want to do with Elasticsearch Connector. I struggled with this for hours before I realized that the default help text on this module was entirely not true for Lando. More explicit documentation would be helpful.