Fixes #2934
In c1298b1b6444e0ce627e005484c750c53ac534cb we introduced a change to the laemp proxy config to allow for the Acquia recipe to inject proxy entries Mailhog. This caused standard Drupal recipes to not inject their own proxy. Other Laempy recipes did not seem to be affected in testing.